# Remove dummy root for query execution plan