## Summary

Current query plan shows a dummy root row (`.`) which is rendered by treeprint. This root row is not needed and it's better to treat the first query execution operator as a root row.

This Pull Request has also the following changes.

* Remove `QueryPlanRow.TextOnly` field as the dummy root is removed
* Handle error properly which could happen in building the query execution plan tree.

## Before

```
$ spanner-cli -e 'EXPLAIN ANALYZE SELECT s.SongName, s.Duration FROM Songs@{force_index=SongsBySongName} AS s WHERE STARTS_WITH(s.SongName, "B");' -t
+-----+----------------------------------------------------------------+---------------+------------+---------------+                                                                      
| ID  | Query_Execution_Plan                                           | Rows_Returned | Executions | Total_Latency |
+-----+----------------------------------------------------------------+---------------+------------+---------------+                                                                                                                                                                                        
|     | .                                                              |               |            |               |
|  *0 | +- Distributed Union                                           | 1             | 1          | 0.25 msecs    |
|  *1 |     +- Distributed Cross Apply                                 | 1             | 1          | 0.2 msecs     |
|   2 |         +- [Input] Create Batch                                |               |            |               |                                                                    
|   3 |         |   +- Local Distributed Union                         | 1             | 1          | 0.11 msecs    |                                                                                      
|   4 |         |       +- Compute Struct                              | 1             | 1          | 0.1 msecs     |
|  *5 |         |           +- FilterScan                              | 1             | 1          | 0.1 msecs     |           
|   6 |         |               +- Index Scan (Index: SongsBySongName) | 1             | 1          | 0.09 msecs    |
|  20 |         +- [Map] Serialize Result                              | 1             | 1          | 0.06 msecs    |                          
|  21 |             +- Cross Apply                                     | 1             | 1          | 0.05 msecs    |                                                                                                                                                                                        
|  22 |                 +- [Input] Batch Scan (Batch: $v2)             | 1             | 1          | 0 msecs       |                                                                    
|  27 |                 +- [Map] Local Distributed Union               | 1             | 1          | 0.05 msecs    |                                                                                      
| *28 |                     +- FilterScan                              | 1             | 1          | 0.04 msecs    |                                                                    
|  29 |                         +- Table Scan (Table: Songs)           | 1             | 1          | 0.04 msecs    |
+-----+----------------------------------------------------------------+---------------+------------+---------------+
```

## After

```
$ spanner-cli -e 'EXPLAIN ANALYZE SELECT s.SongName, s.Duration FROM Songs@{force_index=SongsBySongName} AS s WHERE STARTS_WITH(s.SongName, "B");' -t
+-----+------------------------------------------------------------+---------------+------------+---------------+    
| ID  | Query_Execution_Plan                                       | Rows_Returned | Executions | Total_Latency |                                                                                                                                                                                            
+-----+------------------------------------------------------------+---------------+------------+---------------+    
|  *0 | Distributed Union                                          | 1             | 1          | 0.19 msecs    |    
|  *1 | +- Distributed Cross Apply                                 | 1             | 1          | 0.17 msecs    |                                                                        
|   2 |     +- [Input] Create Batch                                |               |            |               |                                                                                          
|   3 |     |   +- Local Distributed Union                         | 1             | 1          | 0.09 msecs    |    
|   4 |     |       +- Compute Struct                              | 1             | 1          | 0.08 msecs    |    
|  *5 |     |           +- FilterScan                              | 1             | 1          | 0.08 msecs    |                              
|   6 |     |               +- Index Scan (Index: SongsBySongName) | 1             | 1          | 0.08 msecs    |                                                                                                                                                                                            
|  20 |     +- [Map] Serialize Result                              | 1             | 1          | 0.04 msecs    |
|  21 |         +- Cross Apply                                     | 1             | 1          | 0.04 msecs    |
|  22 |             +- [Input] Batch Scan (Batch: $v2)             | 1             | 1          | 0 msecs       |
|  27 |             +- [Map] Local Distributed Union               | 1             | 1          | 0.03 msecs    |
| *28 |                 +- FilterScan                              | 1             | 1          | 0.03 msecs    |
|  29 |                     +- Table Scan (Table: Songs)           | 1             | 1          | 0.03 msecs    |
+-----+------------------------------------------------------------+---------------+------------+---------------+
```